Initial commit
This is the first commit, though the project has been worked on for a
few weeks. It currently is at the point of support most route points and
a basic homepage template file. It supports proper sessions with logging
in and out. Since it uses the dialog element, it is mostly only
compatible with Chrome, though it should work in all HTML5 browsers.